completed part 4 with all tests

+++ backend/controllers/blogs.js
@@ -1,6 +1,5 @@
 const blogsRouter = require('express').Router()
 const Blog = require('../models/blog')
-const User = require('../models/user')
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
@@ -9,19 +8,22 @@ blogsRouter.get('/', async (request, response) => {
 
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
 
-  const user = await User.findById(body.userId)
+  if (!user) {
+    response.status(401).json({ error: 'token not provided' })
+  }
 
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes || 0,
-    user: user.id
+    user: user._id,
   })
 
   const savedBlog = await blog.save()
-  user.blogs = user.notes.concat(savedBlog._id)
+  user.blogs = user.blogs.concat(savedBlog._id)
   await user.save()
 
   response.status(201).json(savedBlog)
@@ -38,24 +40,57 @@ blogsRouter.get('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (!blog) {
+    response.status(404).json({ error: 'blog not found' })
+  }
+
+  if (!user) {
+    response.status(401).json({ error: 'token not provided' })
+  }
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({ error: 'token invalid' })
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
   const body = request.body
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
 
-  const blog = {
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
+  if (!blog) {
+    response.status(404).json({ error: 'blog not found' })
   }
 
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {
-    new: true,
-  })
-  response.json(updatedBlog)
+  if (!user) {
+    response.status(401).json({ error: 'token not provided' })
+  }
+
+  if (blog.user.toString() === user._id.toString()) {
+    const updatedBlog = {
+      title: body.title,
+      author: body.author,
+      url: body.url,
+      likes: body.likes || 0,
+    }
+
+    const returnedBlog = await Blog.findByIdAndUpdate(
+      request.params.id,
+      updatedBlog,
+      {
+        new: true,
+      }
+    )
+    response.json(returnedBlog)
+  } else {
+    response.status(401).json({ error: 'token invalid' })
+  }
 })
 
 module.exports = blogsRouter
